requirements.txt
pandas
numpy
scikit-learn
torch
torchvision
tqdm
umap-learn
matplotlib
seaborn
